add initial sample
add first version

fix scalings

add more rendering

rework structure

add undo/redo

add brush tool

add arrow gl

add rect gl

add text rendering

results

add text rendering

add marker

add crop

add blurring

remove renderer

remove draw functions

clean

refactor

clean

add basic rendering support

add scale_factor

add saving and clipboard

fix crash

clean

refacto

clean

clean up, refactor

refactor

refactor

improve handle sizing

rework scale

remove pangocairo

calc everything in f32

refactor more

refactor

celan up

+++ src/command_line.rs
@@ -34,7 +34,7 @@ pub struct CommandLine {
 
     /// Increase or decrease the size of the annotations
     #[arg(long)]
-    pub annotation_size_factor: Option<f64>,
+    pub annotation_size_factor: Option<f32>,
 
     /// After copying the screenshot, save it to a file as well
     #[arg(long)]
